fix: hilt dep issue

+++ app/src/main/java/com/gnexus/app/data/db/GameEntity.kt
@@ -12,7 +12,7 @@ data class GameEntity(
     val coverUrl: String,
     val isOwned: Boolean,
     val playTime: Int,          // Use Duration for time measurement
-    val progress: Float,             // Ensure this is validated to be between 0.0 and 1.0
+    val progress: Double,             // Ensure this is validated to be between 0.0 and 1.0
     val platform: Int,          // Suggestion 2: Use the Platform enum
     val trophyCount: String,    // Suggestion 1 & 4: Use the new class and a clearer name
     val lastPlayedAt: Int        // Suggestion 5: Use Instant for a point in time
